OPENEUROPA-1616: Fixes after review.
