+++ resources/NotificationServiceWSDL.xml
@@ -8,7 +8,7 @@
     <types>
         <xsd:schema>
             <xsd:import namespace="http://eu.europa.ec.dgt.epoetry"
-                        schemaLocation="NotificationService_xsd.xml"/>
+                        schemaLocation="NotificationServiceXSD.xml"/>
         </xsd:schema>
     </types>
     <message name="receiveNotification">
